// src/app/api/models/inventory-dashboard-metrics-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TrendsDto } from '../models/trends-dto';
export interface InventoryDashboardMetricsDto {
  lowStockProducts: number;
  movementsToday: number;
  outOfStockProducts: number;
  overStockProducts: number;
  totalInventoryValue: number;
  totalProducts: number;
  trends: TrendsDto;
}